requirements.txt
numpy
pyyaml
torch
tqdm
scipy
matplotlib
scikit-image
pandas